+++ src/features/chat/components/chat-input/context-settings-menu.tsx
@@ -25,7 +25,11 @@ import { useOpenTabs } from "@/features/tabs/hooks/use-open-tab"
 import { useTabContents } from "@/features/tabs/hooks/use-tab-contents"
 import { useTabStatusMap } from "@/features/tabs/hooks/use-tab-status-map"
 import { useSelectedTabs } from "@/features/tabs/stores/selected-tabs-store"
-import { DEFAULT_EXCLUDE_URLS, STORAGE_KEYS } from "@/lib/constants"
+import {
+  DEFAULT_EXCLUDE_URLS,
+  DEFAULT_TABS_ACCESS,
+  STORAGE_KEYS
+} from "@/lib/constants"
 import { Layers } from "@/lib/lucide-icon"
 import { plasmoGlobalStorage } from "@/lib/plasmo-global-storage"
 import { cn } from "@/lib/utils"
@@ -186,7 +190,7 @@ export const ContextSettingsMenu = () => {
   )
   const [tabAccess, setTabAccess] = useStorage<boolean>(
     { key: STORAGE_KEYS.BROWSER.TABS_ACCESS, instance: plasmoGlobalStorage },
-    false
+    DEFAULT_TABS_ACCESS
   )
   const { tabs: openTabs, refreshTabs } = useOpenTabs(Boolean(tabAccess))
   const { selectedTabIds, setSelectedTabIds } = useSelectedTabs()